trunk: Get Jenkins to run SG++ again. * Fix Python errors by commenting imports (maybe Valeriy forgot to add files). [datadriven] * Fix GCC warnings. [base]
git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/trunk@5487 4eea3252-f0fb-4393-894d-40516dce545b

+++ base/src/sgpp/base/grid/common/IndexInSubspaceGenerator.hpp
@@ -113,7 +113,7 @@ private:
     value_type level_vector;
     value_type max_index_vector;
     pointer_type val_;
-    unsigned int dim_;
+    size_t dim_;
     std::queue <queue_value_type> queue_;
 
     /**
